WORKSPACE/requirements.txt
pandas
numpy
matplotlib
scikit-learn
python-box
pyYAML
ensure
joblib
types-pyYAML
Flask
Flask-Cors
-e .
